Backend/pyproject.toml
[project]
name = "acro"
version = "0.1.0"
description = "ACRO (Agentic Cyber-Orchestration) mission control using CAMEL AI, OWL, and SETA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.30.6",
    "websockets>=13.0",
    "camel-ai>=0.1.0",
    "litellm>=1.46.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "duckduckgo-search>=6.0.0",
    "pytest>=9.0.2",
    "GitPython>=3.1.43",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "features"]
